update examples and readme
